Fix the NULL bug
When apriltag detection is running, if the tag or camera shake
(like visual servoing) sometime the tag cannot be detected and cause the /tag_detection topic fail to publish geometry_msgs::Poses and cause the node
segematation fail.
Therefore, by adding a checker to check the pointer is NULL or not
could prevent this happen.